+++ src/plugins/templatePlugin.ts
@@ -8,6 +8,8 @@ export interface PluginUtilities {
     xmlParser: XmlParser;
 }
 
+/* eslint-disable @typescript-eslint/member-ordering */
+
 export abstract class TemplatePlugin {
 
     public abstract get prefixes(): TagPrefix[];