+++ test/main.ts
@@ -1,3 +1,0 @@
-var ctx = (require as any).context('.', true, /\.(ts|js)$/);
-ctx.keys().forEach(ctx);
-module.exports = ctx;