+++ test/xmlNode.tests.ts
@@ -1,9 +1,6 @@
-import { expect } from 'chai';
-import { XmlNode, XmlNodeType } from 'src/xmlNode';
+import { TEXT_NODE_NAME, XmlNode, XmlNodeType } from 'src/xmlNode';
 import { XmlParser } from 'src/xmlParser';
 import { parseXml } from './testUtils';
-
-// tslint:disable:no-unused-expression
 
 describe(nameof(XmlNode), () => {
 
@@ -12,19 +9,19 @@ describe(nameof(XmlNode), () => {
         it('serializes a simple text node', () => {
             const node = XmlNode.createTextNode('hello');
             const str = XmlNode.serialize(node);
-            expect(str).to.eql('hello');
+            expect(str).toEqual('hello');
         });
 
         it('serializes a text node with null value as an empty string', () => {
             const node = XmlNode.createTextNode(null);
             const str = XmlNode.serialize(node);
-            expect(str).to.eql('');
+            expect(str).toEqual('');
         });
 
         it('serializes a text node with undefined value as an empty string', () => {
             const node = XmlNode.createTextNode(undefined);
             const str = XmlNode.serialize(node);
-            expect(str).to.eql('');
+            expect(str).toEqual('');
         });
 
     });
@@ -36,11 +33,11 @@ describe(nameof(XmlNode), () => {
             const domNode = createDomNode('<my-node/>');
             const xmlNode = XmlNode.fromDomNode(domNode);
 
-            expect(xmlNode.nodeName).to.eql('my-node');
-            expect(xmlNode.nodeType).to.eql(XmlNodeType.General);
-            expect(xmlNode.parentNode).to.not.exist;
-            expect(xmlNode.childNodes.length).to.eql(0);
-            expect(xmlNode.nextSibling).to.not.exist;
+            expect(xmlNode.nodeName).toEqual('my-node');
+            expect(xmlNode.nodeType).toEqual(XmlNodeType.General);
+            expect(xmlNode.parentNode).toBeFalsy();
+            expect(xmlNode.childNodes.length).toEqual(0);
+            expect(xmlNode.nextSibling).toBeFalsy();
         });
 
         it('creates a valid xml tree from a dom node with a single child', () => {
@@ -53,19 +50,19 @@ describe(nameof(XmlNode), () => {
 
             // root
             const root = xmlNode;
-            expect(root.nodeName).to.eql('root');
-            expect(xmlNode.nodeType).to.eql(XmlNodeType.General);
-            expect(root.parentNode).to.not.exist;
-            expect(root.childNodes.length).to.eql(1);
-            expect(root.nextSibling).to.not.exist;
+            expect(root.nodeName).toEqual('root');
+            expect(xmlNode.nodeType).toEqual(XmlNodeType.General);
+            expect(root.parentNode).toBeFalsy();
+            expect(root.childNodes.length).toEqual(1);
+            expect(root.nextSibling).toBeFalsy();
 
             // child
             const child = root.childNodes[0];
-            expect(child.nodeName).to.eql('child');
-            expect(xmlNode.nodeType).to.eql(XmlNodeType.General);
-            expect(child.parentNode).to.eql(root);
-            expect(child.childNodes.length).to.eql(0);
-            expect(child.nextSibling).to.not.exist;
+            expect(child.nodeName).toEqual('child');
+            expect(xmlNode.nodeType).toEqual(XmlNodeType.General);
+            expect(child.parentNode).toEqual(root);
+            expect(child.childNodes.length).toEqual(0);
+            expect(child.nextSibling).toBeFalsy();
         });
 
         it('creates a valid xml tree from a mixed tree', () => {
@@ -80,11 +77,11 @@ describe(nameof(XmlNode), () => {
 
             // root
             const root = xmlNode;
-            expect(root.nodeName).to.eql('root');
-            expect(root.nodeType).to.eql(XmlNodeType.General);
-            expect(root.parentNode).to.not.exist;
-            expect(root.childNodes.length).to.eql(3);
-            expect(root.nextSibling).to.not.exist;
+            expect(root.nodeName).toEqual('root');
+            expect(root.nodeType).toEqual(XmlNodeType.General);
+            expect(root.parentNode).toBeFalsy();
+            expect(root.childNodes.length).toEqual(3);
+            expect(root.nextSibling).toBeFalsy();
 
             const child1 = root.childNodes[0];
             const child2 = root.childNodes[1];
@@ -92,32 +89,32 @@ describe(nameof(XmlNode), () => {
             const grandchild1 = root.childNodes[2].childNodes[0];
             
             // child #1
-            expect(child1.nodeName).to.eql('child');
-            expect(child1.nodeType).to.eql(XmlNodeType.General);
-            expect(child1.parentNode).to.eql(root);
-            expect(child1.childNodes.length).to.eql(0);
-            expect(child1.nextSibling).to.eql(child2);
+            expect(child1.nodeName).toEqual('child');
+            expect(child1.nodeType).toEqual(XmlNodeType.General);
+            expect(child1.parentNode).toEqual(root);
+            expect(child1.childNodes.length).toEqual(0);
+            expect(child1.nextSibling).toEqual(child2);
 
             // child #2
-            expect(child2.nodeName).to.eql('child');
-            expect(child2.nodeType).to.eql(XmlNodeType.General);
-            expect(child2.parentNode).to.eql(root);
-            expect(child2.childNodes.length).to.eql(0);
-            expect(child2.nextSibling).to.eql(child3);
+            expect(child2.nodeName).toEqual('child');
+            expect(child2.nodeType).toEqual(XmlNodeType.General);
+            expect(child2.parentNode).toEqual(root);
+            expect(child2.childNodes.length).toEqual(0);
+            expect(child2.nextSibling).toEqual(child3);
 
             // child #3
-            expect(child3.nodeName).to.eql('other-child');
-            expect(child3.nodeType).to.eql(XmlNodeType.General);
-            expect(child3.parentNode).to.eql(root);
-            expect(child3.childNodes.length).to.eql(1);
-            expect(child3.nextSibling).to.not.exist;
+            expect(child3.nodeName).toEqual('other-child');
+            expect(child3.nodeType).toEqual(XmlNodeType.General);
+            expect(child3.parentNode).toEqual(root);
+            expect(child3.childNodes.length).toEqual(1);
+            expect(child3.nextSibling).toBeFalsy();
 
             // grandchild #3
-            expect(grandchild1.nodeName).to.eql(XmlNode.TEXT_NODE_NAME);
-            expect(grandchild1.nodeType).to.eql(XmlNodeType.Text);
-            expect(grandchild1.parentNode).to.eql(child3);
-            expect(grandchild1.childNodes).to.not.exist;
-            expect(grandchild1.nextSibling).to.not.exist;
+            expect(grandchild1.nodeName).toEqual(TEXT_NODE_NAME);
+            expect(grandchild1.nodeType).toEqual(XmlNodeType.Text);
+            expect(grandchild1.parentNode).toEqual(child3);
+            expect(grandchild1.childNodes).toBeFalsy();
+            expect(grandchild1.nextSibling).toBeFalsy();
         });
 
     });
@@ -141,21 +138,21 @@ describe(nameof(XmlNode), () => {
             XmlNode.insertBefore(newChild, child);
 
             // parent
-            expect(parent.parentNode).to.not.exist;
-            expect(parent.nextSibling).to.not.exist;
-            expect(parent.childNodes.length).to.eql(2);
-            expect(parent.childNodes[0]).to.eql(newChild);
-            expect(parent.childNodes[1]).to.eql(child);
+            expect(parent.parentNode).toBeFalsy();
+            expect(parent.nextSibling).toBeFalsy();
+            expect(parent.childNodes.length).toEqual(2);
+            expect(parent.childNodes[0]).toEqual(newChild);
+            expect(parent.childNodes[1]).toEqual(child);
 
             // child
-            expect(child.parentNode).to.eql(parent);
-            expect(child.childNodes.length).to.eql(0);
-            expect(child.nextSibling).to.not.exist;
+            expect(child.parentNode).toEqual(parent);
+            expect(child.childNodes.length).toEqual(0);
+            expect(child.nextSibling).toBeFalsy();
 
             // new child
-            expect(newChild.parentNode).to.eql(parent);
-            expect((newChild.childNodes || []).length).to.eql(0);
-            expect(newChild.nextSibling).to.eql(child);
+            expect(newChild.parentNode).toEqual(parent);
+            expect((newChild.childNodes || []).length).toEqual(0);
+            expect(newChild.nextSibling).toEqual(child);
         });
 
     });
@@ -174,15 +171,15 @@ describe(nameof(XmlNode), () => {
             XmlNode.insertChild(parent, child, 0);
 
             // parent
-            expect(parent.parentNode).to.not.exist;
-            expect(parent.nextSibling).to.not.exist;
-            expect(parent.childNodes.length).to.eql(1);
-            expect(parent.childNodes[0]).to.eql(child);
+            expect(parent.parentNode).toBeFalsy();
+            expect(parent.nextSibling).toBeFalsy();
+            expect(parent.childNodes.length).toEqual(1);
+            expect(parent.childNodes[0]).toEqual(child);
 
             // child
-            expect(child.parentNode).to.eql(parent);
-            expect((child.childNodes || []).length).to.eql(0);
-            expect(child.nextSibling).to.not.exist;
+            expect(child.parentNode).toEqual(parent);
+            expect((child.childNodes || []).length).toEqual(0);
+            expect(child.nextSibling).toBeFalsy();
 
         });
 
@@ -207,21 +204,21 @@ describe(nameof(XmlNode), () => {
             XmlNode.appendChild(parent, newChild);
 
             // parent
-            expect(parent.parentNode).to.not.exist;
-            expect(parent.nextSibling).to.not.exist;
-            expect(parent.childNodes.length).to.eql(2);
-            expect(parent.childNodes[0]).to.eql(child);
-            expect(parent.childNodes[1]).to.eql(newChild);
+            expect(parent.parentNode).toBeFalsy();
+            expect(parent.nextSibling).toBeFalsy();
+            expect(parent.childNodes.length).toEqual(2);
+            expect(parent.childNodes[0]).toEqual(child);
+            expect(parent.childNodes[1]).toEqual(newChild);
 
             // child
-            expect(child.parentNode).to.eql(parent);
-            expect(child.childNodes.length).to.eql(0);
-            expect(child.nextSibling).to.eql(newChild);
+            expect(child.parentNode).toEqual(parent);
+            expect(child.childNodes.length).toEqual(0);
+            expect(child.nextSibling).toEqual(newChild);
 
             // new child
-            expect(newChild.parentNode).to.eql(parent);
-            expect((newChild.childNodes || []).length).to.eql(0);
-            expect(newChild.nextSibling).to.not.exist;
+            expect(newChild.parentNode).toEqual(parent);
+            expect((newChild.childNodes || []).length).toEqual(0);
+            expect(newChild.nextSibling).toBeFalsy();
 
         });